Split component - Credential section is now separate - Will be useful when other sections like projects and blogs will be added - Modular code
